requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
google-generativeai
groq
langchain-google-genai
langchain-groq
faiss-cpu
pymupdf
pandas
python-dotenv
boto3
sentence-transformers
tiktoken